feat: convert AUpload to typescript and add more docs
