+++ framework/components/AUpload/AUpload.mdx
@@ -8,6 +8,8 @@ sourceCodeLink: https://github.com/cisco-sbg-ui/magna-react/tree/main/framework/
 
 ## Usage
 
+AUpload uses [`react-dropzone`](`https://react-dropzone.js.org/`) internally to manage its state. `react-dropzone` props can be passed to `AUpload` via the `dropzoneProps` prop. Please check out their docs for more information.
+
 ### Import
 
 <ComponentImport components="AUpload" />
